% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statistics_functions.R
\name{afl}
\alias{afl}
\title{Average future lifetime}
\usage{
afl(init_age, init_state, trans_probs)
}
\arguments{
\item{init_age}{integer between 65 and 110 denoting initial age of individual}

\item{init_state}{0 for healthy, 1 for disabled}

\item{trans_probs}{a list of transition probability matrices, preferably generated from
\code{\link[tshm]{get_trans_probs}}.}
}
\value{
numeric output for average and standard deviation of future lifetime
}
\description{
Calculates the average future life time and its standard deviation
given initial state and age of an individual by simulating life time paths.
NOTE: USE \code{\link[tshm]{aflF}} for frailty model.
}
