% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statistics_functions.R
\name{afld}
\alias{afld}
\title{Average future lifetime in disabled state}
\usage{
afld(init_age, init_state, trans_probs, n = 5e+05)
}
\arguments{
\item{init_age}{Integer between 65 and 110 specifying the initial age of the individual}

\item{init_state}{0 for healthy, 1 for disabled}

\item{trans_probs}{list of transition probability matrices, preferably generated from \code{\link[tshm]{get_trans_probs}}.}

\item{n}{integer denoting number of individuals in simulation}
}
\value{
Mean and standard deviation of average disabled times across n simulations.
}
\description{
Calculates the average future lifetime that will be spent in disabled state from
initial age and state. This average future lifetime depends on the initial state
the individual is in. Generally, this is applied to non-homogeneus Markov chains,
so we will use stochastic simulation to calculate this statistic.
NOTE: THIS FUNCTION SHOULD ONLY BE USED WITH STATIC AND TREND MODEL, USE ....
FOR FRAILTY MODEL.
}
