% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statistics_functions.R
\name{afldF}
\alias{afldF}
\title{Average future lifetime in disabled state (frailty moded)}
\usage{
afldF(init_age, init_state, female, year, param_file, n = 5000)
}
\arguments{
\item{init_age}{integer between 65 and 110 denoting initial age of individual}

\item{init_state}{0 for healthy and 1 for disabled}

\item{female}{0 for non female, 1 for female}

\item{year}{integer for current year}

\item{param_file}{string for file path containing parameters of cox regression model}

\item{n}{integer representing number of unique latent factors to simulate}
}
\value{
numeric output for average time spent in disabled state
}
\description{
Performs the same function as \code{\link[tshm]{afld}}, but it requires inputs
to simulate latent factor in the frailty model. This allows it to simulate the
the randomness from the frailty factor into the calculation as well.
NOTE: THIS FUNCTION SHOULD ONLY BE USED WITH THE FRAILTY MODEL. USE \code{\link[tshm]{afld}}
FOR STATIC AND TREND MODEL.
}
