% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statistics_functions.R
\name{aflF}
\alias{aflF}
\title{Average future lifetime (Frailty Model)}
\usage{
aflF(init_age, init_state, female, year, param_file, n = 5000)
}
\arguments{
\item{init_age}{integer between 65 and 110 indicating initial age of individual}

\item{init_state}{0 for healthy, 1 for disabled}

\item{female}{0 for male, 1 for female}

\item{year}{integer indicating current year}

\item{param_file}{string name of file containing all the parameters of the model}

\item{n}{integer denoting number of unique latent factor simulations}
}
\value{
numeric output for expected future lifetime
}
\description{
Performs the same function as the \code{\link[tshm]{afl}}, but simulates a number of
unique latent factor paths to make probabilities stochastic.
NOTE: USE \code{\link[tshm]{afl}} for static and trend models.
}
