% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{china_CLHLS}
\alias{china_CLHLS}
\title{Parameters of Cox Regression Model estimated from China CLHLS Study}
\format{
A tibble OR data frame with 15 rows and 5 columns:
\describe{
\item{Model}{static, trend or frailty}
\item{Coef}{the covariate that the coefficient is corresponding to}
\item{trans1}{parameter estimate for H to F transition}
\item{trans2}{parameter estimate for F to H transition}
\item{trans3}{parameter estimate for H to D transition}
\item{trans4}{parameter estimate for F to D transition}
}
}
\source{
\url{https://cepar.edu.au/publications/working-papers/functional-disability-systematic-trends-and-uncertainty-comparison-between-china-and-us}
}
\usage{
china_CLHLS
}
\description{
A dataset containing the parameters of the static, trend and frailty cox regression
models estimated in the 1998-2014 Chinese Longitudinal Healthy Longevity Survey (CLHLS). The
model includes parameters for the following covariates: age, gender, time (trend) stochastic
latent factor (frailty).
}
\keyword{datasets}
