% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_life_table.R
\name{create_life_table}
\alias{create_life_table}
\title{Create life table}
\usage{
create_life_table(trans_probs, age, init_state = 0, cohort = 1e+05)
}
\arguments{
\item{trans_probs}{list of transition probability matrices; typically generated by \code{\link[tshm]{get_trans_probs}}}

\item{age}{integer denoting current age}

\item{init_state}{integer value of 0 or 1, where 0 indicates healthy and 1 indicates disabled}

\item{cohort}{initial cohort size for lifetable}
}
\value{
dataframe of lifetable generated
}
\description{
Creates life table from a list of transition probability matrices at some initial age
and specified cohort size.
}
\examples{

}
