% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statistics_functions.R
\name{hfl}
\alias{hfl}
\title{Healthy Future lifetime}
\usage{
hfl(init_age, init_state, trans_probs)
}
\arguments{
\item{init_age}{integer between 65 and 110 denoting initial age of individual}

\item{init_state}{0 for healthy, 1 for disabled}

\item{trans_probs}{list of transition probability matrices; ideally generated from
\code{\link[tshm]{get_trans_probs}}.}
}
\value{
Numeric output of expected time and standard deviation of time spent in healthy state.
}
\description{
Calculates the expected future lifetime spent and its standard deviation
in the healthy state.
NOTE: USE \code{\link[tshm]{hflF}} FOR FRAILTY MODEL.
}
