% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots_functions.R
\name{prob_plots}
\alias{prob_plots}
\title{Create Survival Probability Plots}
\usage{
prob_plots(init_state, init_age, trans_probs)
}
\arguments{
\item{init_state}{0 for healthy, 1 for disabled.}

\item{init_age}{integer between 65 and 110 (inclusive) denoting initial age of individual. This has
to be the same as the initial age used for transition probability matrices simulation.}

\item{trans_probs}{list of transition probability matrices, this could be generated by \code{\link[tshm]{get_trans_probs}}.}
}
\value{
line plot with 3 lines, each denoting survival probability, survivial probability to
disabled, and survival probability to healthy.
}
\description{
Create a plot of survival probabilities to alive, health, and disabled states.
}
