% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_path.R
\name{simulate_path}
\alias{simulate_path}
\title{Simulate cohort life path}
\usage{
simulate_path(init_age, init_state, cohort = 10000, trans_probs)
}
\arguments{
\item{init_age}{integer between 64 and 110 denoting current age}

\item{init_state}{1 for healthy, 2 for disabled}

\item{cohort}{integer (default 10000) denoting number of people in the simulation}

\item{trans_probs}{a list of transition probability matrices, preferably generated from \code{\link[tshm]{get_trans_probs}}.}
}
\value{
a matrix where each row represents a new individual, and the columns represent
the individual's movement through each state.

3 (death) is absorbing, so if an individual enters that state, the rest of the row will be 3.
}
\description{
Simulates the path each life takes in an initial cohort using transition probabilities
}
