% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statistics_functions.R
\name{time_to_disabled}
\alias{time_to_disabled}
\title{Time until onset of disability (conditional on being disabled)}
\usage{
time_to_disabled(init_age, trans_probs)
}
\arguments{
\item{init_age}{integer between 65 and 110 denoting initial age of individual}

\item{trans_probs}{a list of transition probability matrices, preferably generated from
\code{\link[tshm]{get_trans_probs}}.}
}
\value{
numeric denoting the average time until individual becomes disabled and its
standard deviation.
}
\description{
Uses simulation to produce an average time until a healthy individual becomes disabled
during their life time. Note that this is conditional on the individual becoming disabled,
so it doesn't factor in life simulations where the indiviudal is always healthy.
NOTE: USE \code{\link[tshm]{time_to_disabledF}} for frailty model.
}
