% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statistics_functions.R
\name{afl}
\alias{afl}
\title{average future lifetime}
\usage{
afl(life_table)
}
\arguments{
\item{life_table}{This is a life table generated from \code{\link[tshm]{create_life_table}}.
Since the creation of the life table already has an initial state that the cohort
is in, the function does not require the same input.}
}
\value{
Numeric output for the average future lifetime. This is calculated by taking the
curtate expected lifetime; essentially a sum of survival probabilities.
}
\description{
Calculates average future lifetime remaining from initial age and state. Note that the
average future lifetime calculates the remaining time left (regardless of state) from the
initial state. So remaining lifetimes for individuals starting healthy or disabled will differ.
}
