% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statistics_functions.R
\name{hfl}
\alias{hfl}
\title{Healthy Future lifetime}
\usage{
hfl(init_age, init_state, trans_probs = NULL, simulated_path = NULL)
}
\arguments{
\item{init_age}{integer between 65 and 110 denoting initial age of individual}

\item{init_state}{0 for healthy, 1 for disabled}

\item{trans_probs}{list of transition probability matrices; ideally generated from
\code{\link[tshm]{get_trans_probs}}.}

\item{simulated_path}{a matrix of lifetime paths simulations, generated by \code{\link[tshm]{simulate_path}}.}
}
\value{
Numeric output of expected time and standard deviation of time spent in healthy state.
}
\description{
Calculates the expected future lifetime spent and its standard deviation
in the healthy state. Function uses either transition probability matrices, or
a simulated lifetime paths. If both are provided, it will use the simulated lifetime
provided.
NOTE: USE \code{\link[tshm]{hflF}} FOR FRAILTY MODEL.
}
